Day 0/requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.19.4
pandas==1.2.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
six==1.15.0
urllib3==1.26.2
